+++ angular_test/CHANGELOG.md
@@ -1,5 +1,7 @@
 *   Deleted an unnecessary `hostElement.append(componentRef.location)`.
 
+*   Supported `FutureOr<void>` for `beforeChangeDetection`.
+
 ## 2.0.0-beta+2
 
 *   Maintanence release: Stop using `SCREAMING_CAPS` consts from `dart:*`.